feat(sample): add `_meta` sections to homebrew files
This demonstrates a simple `_meta` addon which has the "important bits"
of the standard homebrew `_meta`, in a slightly-flattened format (this
is due to each homebrew file in this repo matching exactly one homebrew
`"json"` source, whereas a file in the homebrew repo can contain
multiple `"json"` sources).

We choose to omit the `author`, `url`, etc. properties, as this
information is already stored in the homebrew repo, and duplicating it
has no benefit.