feat(effects): explicitly set disabled/transfer values
Plutonium's AE handling, when it comes to compatibility with
MidiQOL/DAE, is currently abysmal. As part of efforts to straighten that
out moving forward, be more explicit about which effects have which
`transfer`/`disabled` values.

Additionally, avoid using MidiQOl-specific versions of these flags, as
they don't seem to take precedence over the real `transfer`/`disabled`,
which Plutonium always sets to a non-null value internally.

Note that for all "addon" effects, the implicit default for both values
is `false`. This also applies to the `foundry-*.json`s/etc. in the
5etools source, so copy-pasting between the two won't require any
changes.