doc(README): list Item Macro as hard dependency; use package page links
Revise docs to list Item Macro as a hard dependency, as we have already
baked in both the dependency as well as macros which use it.

Additionally, prefer Foundry package links over e.g. module repo links
in the "user-facing" README, to improve non-dev UX and make our links
less brittle in the face of e.g. a module author moving their code.

Closes #72