fix(integration): allow CPR doc diff against partial base content
Apply the base system data to the faux item document we pass to CPR;
we do this so that CPR's changes are made against a reasonable
approximation of the to-be-created document, rather than an empty
document. This avoids cases where e.g. CPR fails to wipe the `damage`
from a spell which relies on a macro for its damage roll
(see: Scorching Ray).