feat(spell): add DDBI automations/macros
Here we fill in gaps left by CPR/MidiSRD, by grabbing available
automations from the D&D Beyond Importer module. We do this, instead of
depending on the module directly, for two reasons:

- DDBI, in code terms, is structured in such a way as to make pulling
  these automations out at runtime difficult. Viable options are limited
  to exceedingly flaky/dangerous big-`eval`s, or, having the user import
  everything from D&D Beyond in advance ( :^) ).
- DDBI represents a, to put it mildly, alternate workflow for importing
  content. Forcing the user into running both is bloaty at best.

As DDBI is regularly updated, keeping especially the macros in-sync is a
task which will require future maintenance. Hopefully as other
integrations advance, some of these spells will be automated elsewhere,
and we can prune down the list of copies we're maintaining. In the
meantime, additions have been made to the converter page to facilitate
easy export of `system` data, and of item macros*.

In particular, some DDBI macros make use of a DDBI module API; a 1:1
port of this has therefore been made here (with renames to avoid any
current and future collisions). As the references to this API are
minimal and easy to find-replace, this solution seems "good enough" for
the moment.

Regardless of potential future maintenance, various spells have seen
basic automation improvements which should make for an all-round better
experience moving forward.

* A future migration should likely be done to transition towards the new
  DAE "DIME" system, but this is work for another PR.