fix(integration): ensure CPR usability for players during import
Either CPR's "Item Configuration Access" option or
"Item Replacer Access" option needs to be enabled, to allow the
integration to grab the appropriate hook. Without this, the integration
does not work for non-GM users.

As "Item Configuration Access" is the lesser of these two options (it
allows players to e.g. adjust the animations, damage types, etc. of
their CPR items), only require that it be enabled, rather than full
"Item Replacer Access" (which has a greater potential for misuse).

Then, during our CPR-data-fetching step, add a player-only patch to
"enable" the "Item Replacer Access" option, such that we can update our
temp items. In practical terms, this is either done for too short an
interval for the player to notice, or is done while wider activity has
probably locked up the main thread and is preventing any user
interaction anyway, so shouldn't present much opportunity for damage
(intentional or otherwise [the same "intentional" trick could be
pulled much easier with a console one-liner!]).