chore: remove DDBI spell clones in favor of CPR versions
Remove the need to maintain the following spells:

- Branding Smite
- Elemental Weapon
- Heroism
- Melf's Acid Arrow
- Spirit Guardians

See previous commit: 91180778.

Note that Spirit Guardians was previously retained as having a better
implementation (with included the "affected creature's speed is halved
in the area" effect) in DDBI; this decision has been reversed, as the
CPR implementation includes a choice of damage types, which the DDBI
version does not*.

Given the choice of maintaining the DDBI version or getting a comparable
CPR version with an alternate feature set, the latter is more
attractive.

* (Due to the way we're abusing it out of context!)